Updated plugin for Transmit 4
* Fixed reading .plists/prefs to get list of favorites
* Switched to Scripting Bridge
* Added a 'mount as disk' action
* Updated documentation
* 64 bit and usual tidyups

+++ src/FavoriteCollection.m
@@ -6,7 +6,7 @@
 
 //static NSImage *genericFolderImage = nil;
 
-int CompareByName(id obj1, id obj2, void* context)
+NSInteger CompareByName(id obj1, id obj2, void* context)
 {
 	NSString*	name1	= nil;
 	NSString*	name2	= nil;
